% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{average_normals}
\alias{average_normals}
\title{Compute mean, upper quantiles, and lower quantiles of average of normal
distributions for each time step.}
\usage{
average_normals(means, stddevs, alpha, drop = FALSE)
}
\arguments{
\item{means}{An \code{nsteps} by \code{ndraws} matrix of means.}

\item{stddevs}{An \code{nsteps} by \code{ndraws} matrix of std devs.}

\item{alpha}{A nonempty vector of values lying strictly between 0 and 0.5.}

\item{drop}{Whether to return \code{lower} and \code{upper} as vectors
instead of matrices when \code{alpha} is scalar.}
}
\value{
List with elements
  \itemize{
  \item \code{mean}: vector of length \code{nsteps}.
  \item \code{lower}: \code{nsteps} by \code{nalpha} matrix.
  \item \code{upper}: \code{nsteps} by \code{nalpha} matrix.
 }
 where \code{mean}[t] is mean of the average of the normal distributions
 for step t, \code{lower}[t,i] is quantile \code{alpha}[i]/2, and
 \code{upper}[t,i] is quantile 1-\code{alpha}[i]/2.
}
\description{
Compute mean, upper quantiles, and lower quantiles of average of normal
distributions for each time step.
}
