% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{averagedLL}
\alias{averagedLL}
\title{Log-likelihood for average of models.}
\usage{
averagedLL(models, y)
}
\arguments{
\item{models}{List of SSMs corresponding to posterior parameter draws}

\item{y}{Sequence of observed values}
}
\value{
Log likelihood of y given average of models.
}
\description{
Log-likelihood for average of models.
}
