% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{cksCompile}
\alias{cksCompile}
\title{Compile a Chronikis program}
\usage{
cksCompile(fpathCKS, fctNameR)
}
\arguments{
\item{fpathCKS}{Path to Chronikis source file.}

\item{fctNameR}{(Optional) Name of R SSM-creation function to generate.
 Defaults to basename of sourcefile with .cks extension removed, prefixed
 with "createSSMs_". In all cases, if the result is not a valid R name
 then it is converted to one in the same way as the `R` function
`make.names`.}
}
\value{
\code{stanmodel} object obtained by compiling the Chronikis program
  to Stan and then running the Stan compiiler.
}
\description{
Compile a Chronikis program
}
