% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{filter_models}
\alias{filter_models}
\title{Apply Kalman filter to each of a list of SSMs.}
\usage{
filter_models(y, models)
}
\arguments{
\item{y}{Time series (ts object or numeric vector) of observations.}

\item{models}{List of SSMs.}
}
\description{
Apply Kalman filter to each of a list of SSMs.
}
