% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{forecast_sample}
\alias{forecast_sample}
\title{Draw a sample from the predictive distribution for the next \code{nsteps}
time steps.}
\usage{
forecast_sample(models, nsteps, ndraws_per_model)
}
\arguments{
\item{models}{List of SSMs corresponding to posterior draws of parameters.}

\item{nsteps}{Number of steps ahead to forecast.}

\item{ndraws_per_model}{Number of draws from predictive distribution for
each SSM in \code{models}.}
}
\value{
\code{nsteps} by \code{nd} matrix of predictive draws,
  \code{nd} = length(\code{models}) * \code{ndraws_per_model}.
}
\description{
Draw a sample from the predictive distribution for the next \code{nsteps}
time steps.
}
