% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{map_estimate}
\alias{map_estimate}
\title{Estimate a Chronikis model using MAP.}
\usage{
map_estimate(y, mdl, draws = 0, ...)
}
\arguments{
\item{y}{A single time series, either as a ts object or a numeric vector.}

\item{mdl}{The model with arguments included, as returned by \code{setArgs} or \code{setArgsList}.}

\item{draws}{(Optional, default 0) If nonzero, the Hessian of the log posterior is computed and this
many draws from the corresponding multivariate normal approximation to the posterior are return.}

\item{...}{Any additional arguments to pass to \code{rstan::optimizing}.}
}
\value{
The object containing the results of estimation. (See \code{rstan::optimizing}.)
}
\description{
Estimate a Chronikis model using MAP.
}
