% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{mdlArgs}
\alias{mdlArgs}
\title{Create arguments to pass to a Chronikis model.}
\usage{
mdlArgs(...)
}
\arguments{
\item{...}{A list of named arguments of the form param_name=param_value, one
for each parameter of the Chronikis model.}
}
\description{
Create arguments to pass to a Chronikis model.
}
