% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm-creation-utils.R
\name{num_draws}
\alias{num_draws}
\title{Number of draws in posterior sample}
\usage{
num_draws(postSample)
}
\arguments{
\item{postSample}{A posterior sample returned by \code{posterior_sample()}.}
}
\description{
Number of draws in posterior sample
}
