% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{posterior_sample}
\alias{posterior_sample}
\title{Extract posterior sample of model parameters from estimation results.}
\usage{
posterior_sample(fit, ndraws = NULL)
}
\arguments{
\item{fit}{A value returned by \code{hmc_estimate()}, \code{vb_estimate()}, or \code{map_estimate()}.}

\item{ndraws}{(optional) number of posterior draws to extract. If not specified,
defaults to the number of draws taken during estimation.}
}
\description{
Extract posterior sample of model parameters from estimation results.
}
