% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{setArgs}
\alias{setArgs}
\title{Add model arguments to compiled model}
\usage{
setArgs(stanmodel, args)
}
\arguments{
\item{stanmodel}{Object returned by one of the \code{_estimate} functions}

\item{args}{Object created using \code{mdlArgs()}, containing the known parameter
settings for the model.}
}
\description{
Add model arguments to compiled model
}
