% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{setArgsList}
\alias{setArgsList}
\title{Combine setArgs and mdlArgs into a single call.}
\usage{
setArgsList(stanmodel, ...)
}
\arguments{
\item{stanmodel}{Object returned by one of the \code{_estimate} functions}

\item{...}{Named arguments param_name=param_value.}
}
\description{
Combine setArgs and mdlArgs into a single call.
}
