% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{smoothed_components}
\alias{smoothed_components}
\title{Create smoothed estimates of SSM components}
\usage{
smoothed_components(y, filtered, Zcomp, drop = FALSE)
}
\arguments{
\item{y}{Observed time series.}

\item{filtered}{Result of \code{filter_models(y, models)}.}

\item{Zcomp}{m-column matrix, where m is length of latent state vector.
Zcomp[i, ] %*% alpha_t selects component i of the SSM at time t when
alpha_t is the latent state at time t.}
}
\value{
List of two leny x nmdl x ncomp arrays \code{means} and \code{stddevs},
  where leny is \code{length(y)}, nmdl is \code{length(filtered)}, and
  ncomp is \code{nrow(Zcomp)}.
  \code{means}[t,k,i] and \code{stddevs}[t,k,i] are the posterior mean and
  standard deviation for component i of model k at time t.
}
\description{
Create smoothed estimates of SSM components
}
