% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{smoothed_ts}
\alias{smoothed_ts}
\title{Create smoothed version of time series.}
\usage{
smoothed_ts(y, filtered)
}
\arguments{
\item{y}{Observed time series.}

\item{filtered}{Result of \code{filter_models(y, models)}.}
}
\value{
List of two leny x nmdl matrices \code{means} and \code{stddevs},
  where leny is \code{length(y)} and nmdl is \code{length(models)}.
  \code{means}[t,k] and \code{stddevs}[t,k] are the posterior mean and
  standard deviation for Z %*% alpha for model k at time t.
}
\description{
Create smoothed version of time series.
}
