% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm-creation-utils.R
\name{diagv}
\alias{diagv}
\title{Create a diagonal matrix from its diagonal.
Utility function used by generated R code.}
\usage{
diagv(v)
}
\arguments{
\item{v}{The desired diagonal vector}
}
\description{
Create a diagonal matrix from its diagonal.
Utility function used by generated R code.
}
